Enum askalono::LicenseType

pub enum LicenseType {
    Original,
    Header,
    Alternate,
}

The type of a license entry (typically in a Store).

Variants
Original

The canonical text of the license.

Header

A license header. There may be more than one in a Store.

Alternate

An alternate form of a license. This is intended to be used for alternate formats of a license, not for variants where the text has different meaning. Not currently used in askalono's SPDX dataset.